updated files

+++ exp_rand.h
@@ -1,6 +1,5 @@
 #ifndef EXPRAND
 #define EXPRAND
-
-double exp_rand(double avg);
+int exp_rand(double avg);
 
 #endif
